+++ tests/test_basic.py
@@ -1,8 +1,19 @@
 import pytest
 
+from ereuse_devicehub.devicehub import Devicehub
 from ereuse_devicehub.client import Client
 
 
+@pytest.mark.mvp
+def test_dummy(_app: Devicehub):
+    """Tests the dummy cli command."""
+    runner = _app.test_cli_runner()
+    runner.invoke('dummy', '--yes')
+    with _app.app_context():
+        _app.db.drop_all()
+
+
+@pytest.mark.mvp
 def test_dependencies():
     with pytest.raises(ImportError):
         # Simplejson has a different signature than stdlib json
@@ -12,6 +23,7 @@ def test_dependencies():
 
 
 # noinspection PyArgumentList
+@pytest.mark.mvp
 def test_api_docs(client: Client):
     """Tests /apidocs correct initialization."""
     docs, _ = client.get('/apidocs')
