--- tests/test_dummy.py
@@ -1,9 +1,0 @@
-from ereuse_devicehub.devicehub import Devicehub
-
-
-def test_dummy(_app: Devicehub):
-    """Tests the dummy cli command."""
-    runner = _app.test_cli_runner()
-    runner.invoke('dummy', '--yes')
-    with _app.app_context():
-        _app.db.drop_all()
